feature: added more Android Architecture Components
- added AlarmsViewModel.java, AlarmViewModelFactory.java and AppExecutors.java in aac package
- updated AlarmDao.java, AlarmViewModel.java and Repository.java
- updated AlarmsAdapter.java
- updated ResultListener.java
- updated Alarm.java, Snooze.java and Vibration.java models
- updated MainActivity.java and NewAlarmActivity.java
- updated strings.xml

+++ app/src/main/java/vkurman/shiftedalarmclock/interfaces/ResultListener.java
@@ -21,5 +21,5 @@ package vkurman.shiftedalarmclock.interfaces;
  * Version 1.0
  */
 public interface ResultListener {
-    void onResultClick(long id);
+    void onResultClick(int id);
 }
